(PC-30603)[PRO] feat: add music and show types options in DetailsForm
