+++ pro/src/screens/IndividualOffer/DetailsScreen/utils.ts
@@ -1,5 +1,7 @@
 import { CategoryResponseModel, SubcategoryResponseModel } from 'apiClient/v1'
+import { showOptionsTree } from 'core/Offers/categoriesSubTypes'
 import { SelectOption } from 'custom_types/form'
+import { DEFAULT_DETAILS_INTITIAL_VALUES } from './constants'
 
 export const buildCategoryOptions = (
   categories: CategoryResponseModel[]
@@ -21,3 +23,24 @@ export const buildSubcategoryOptions = (
         subCategory.categoryId === categoryId
     )
   )
+
+export const getShowSubTypeOptions = (showType: string): SelectOption[] => {
+  if (showType === DEFAULT_DETAILS_INTITIAL_VALUES.showType) {
+    return []
+  }
+
+  const selectedShowTypeChildren = showOptionsTree.find(
+    (showTypeOption) => showTypeOption.code === parseInt(showType)
+  )?.children
+
+  if (!selectedShowTypeChildren) {
+    return []
+  }
+
+  return selectedShowTypeChildren
+    .map((data) => ({
+      value: data.code.toString(),
+      label: data.label,
+    }))
+    .sort((a, b) => a.label.localeCompare(b.label, 'fr'))
+}
